refactor: rename getStarted action to login, update action name and description
- Rename actions/getStarted.ts -> actions/login.ts
- Update action name from 'Get Started - Login' to 'Tailscale Login'
- Simplify description to match the new name
- Update imports in actions/index.ts and init/index.ts
