+++ startos/init/index.ts
@@ -4,7 +4,7 @@ import { setInterfaces } from '../interfaces'
 import { versionGraph } from '../versions'
 import { restoreInit } from '../backups'
 import { actions } from '../actions'
-import { getStarted } from '../actions/getStarted'
+import { getStarted } from '../actions/login'
 import { registerUrlPlugin, exportUrls } from '../plugin/url'
 import { initializeService } from './initializeService'
 import type { Effects } from '@start9labs/start-sdk/base/lib/types'
